+++ apps/angular-lib-dev/src/app/modals/modals.module.ts
@@ -3,13 +3,18 @@ import { CommonModule } from '@angular/common'
 
 import { ModalsRoutingModule } from './modals-routing.module'
 import { ModalsComponent } from './modals.component'
-import { NggDropdownModule, NggModalModule } from '@sebgroup/green-angular'
+import {
+  NggDropdownModule,
+  NggModalModule,
+  NggDatepickerModule,
+} from '@sebgroup/green-angular'
 
 @NgModule({
   declarations: [ModalsComponent],
   imports: [
     CommonModule,
     NggDropdownModule,
+    NggDatepickerModule,
     NggModalModule,
     ModalsRoutingModule,
   ],