+++ libs/core/src/components/index.ts
@@ -12,3 +12,5 @@ export * from '../primitives/listbox/option'
 
 export * from './context-menu/context-menu'
 export * from '../primitives/menu/menu-item'
+
+export * from './datepicker/datepicker'