+++ libs/core/src/primitives/calendar/calendar.styles.ts
@@ -1,0 +1,13 @@
+import { css } from 'lit'
+
+const style = css`
+  td.disabled {
+    color: #999;
+    cursor: default;
+  }
+  td.today {
+    background-color: #eee;
+  }
+`
+
+export default style