+++ libs/core/src/primitives/calendar/calendar.trans.styles.ts
@@ -1,0 +1,8 @@
+import { TransitionalStyles } from '../../utils/helpers/transitional-styles'
+import styles from './calendar.trans.styles.scss'
+
+export function register() {
+  TransitionalStyles.instance.register('gds-calendar', styles.toString())
+}
+
+export default register